only expose rpi core temperature if collected
